apps/backend/requirements.txt
fastapi==0.118.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
uvicorn[standard]==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg[binary]==3.1.18
pytest==8.2.2
httpx==0.27.0
flake8==7.1.1
pytest-cov==5.0.0
